mxc_v4l2_capture: Allow clock setting for de-interlaced video input
Make provision for a decoder driver to specify clock mode for
progressive in case the decoder provides deinterlaced output.

This is required for ADV7280 decoder where we can use the hardware
deinterlace capabity of ADV7280. Without setting the clock mode
to progressive but having hardware deinterlace on for ADV7280
results in a non working setup.

Signed-off-by: Sanchayan Maity <sanchayan.maity@toradex.com>
Acked-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>