ARM: dts: imx7d-colibri: Work around poweroff issue
The AD7879 interrupt GPIO is a wakeup pin. On poweroff, the external
pull-up is not holding the GPIO high, hence causing a falling edge
on the pad. Since the GPIO is a wakeup pin, this causes the SNVS Dumb
PMIC to switch to on again (hence causing a reboot instead of a power-
off).

Make sure the interrupt line stays high during poweroff by configuring
a pull-up on the IOMUXC LPSR GPR.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Max Krummenacher <max.krummenacher@toradex.com>