ARM: dts: imx7-colibri: fix SD-card
The SDHC stack uses the GPIO flags to determine whether the card
detect gpio is active-low or high. Zero means GPIO_ACTIVE_HIGH,
which is actually wrong. Specify the card detect to be active-low.
Also remove the tunning step property, this is a board specific
parameter which seems to come from Sabre, where NXP actually
deleted it too.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Max Krummenacher <max.krummenacher@toradex.com>