sdhci-esdhc-imx: fix bus-width for 1-bit operation.
The 1-bit operation mode, enabled by seeting the 'bus-width' property of
the device tree 'esdhc' node to <1>, not work while using SD card.

The behavior is only noticed when only the data pin 0 is connected to the
hardware. A series of kernel errors are printed to the console, all of them
returning the following error message followed by some explanation:
mmcblk0: error -84 transferring data

If four data lines are connected, it ignores the device-tree
property and works in 4-bit mode of operation without errors. The hardware
used for testing does not support 8-bit mode.

Check the 'bus-width' property and if set to <1>, enable the
SDHCI_QUIRK_FORCE_1_BIT_DATA quirk.

Signed-off-by: Leonardo Graboski Veiga <leonardo.veiga@toradex.com>
Signed-off-by: Stefan Agner <stefan.agner@toradex.com>