ARM: imx: adjust vring location to be at 256MB boundary
The VirtIO vring descriptors are currently at the end of the
1024MiB boundary of DDR memory 0xbfff8000. For the Colibri iMX7S
and iMX7D this is outside of the addressable DDR memory. In
practise VirtIO still worked because when accessing memory at
out-of-bound addresses it usually just accesses memory from lower
addresses (most significant address line is ignored).

However, we should not rely on that behavior. Move the hardcoded
vring to the end of the 256MiB boundary at 0x8fff0000.

Note: This needs a Cortex-M4 firmware which uses the same
hardcoded vring address.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>